Refactor api so that it is clearer which is internal and which isn't
